refactor(plugin): extract host factory and danger lattice from PluginService
- Move createHost, buildProcessApi, buildFsApi, buildGitApi, buildClipboardApi,
  ensureCapabilityConsent, and resolveActiveAgentTerminalId into a new
  electron/services/plugin/PluginHostFactory.ts, threaded through a
  PluginHostFactoryDeps interface holding live references to PluginService's
  shared Maps/Sets/collaborators (never snapshots), matching the existing
  XxxDeps convention used by sibling plugin/ modules
- Move the compound-capability danger lattice (manifestTriggersCompoundElevation,
  computePluginDanger) into electron/services/plugin/pluginDangerLattice.ts
- Add a trackPluginDisposer helper in PluginHostFactory.ts replacing 5 copy-pasted
  get-or-create/push/self-splicing disposer blocks inside the host closures
- PluginService.ts shrinks from 5,512 to 3,789 lines; public exports and
  behavior are unchanged
